Feat: 소속 부서별 요청 확인

+++ src/main/java/bitedu/bipa/simplesignbackend/service/FormManageService.java
@@ -3,11 +3,14 @@ package bitedu.bipa.simplesignbackend.service;
 import bitedu.bipa.simplesignbackend.dao.FormManageDAO;
 import bitedu.bipa.simplesignbackend.mapper.CommonMapper;
 import bitedu.bipa.simplesignbackend.model.dto.*;
+import bitedu.bipa.simplesignbackend.utils.SessionUtils;
 import bitedu.bipa.simplesignbackend.validation.CommonErrorCode;
+import bitedu.bipa.simplesignbackend.validation.CustomErrorCode;
 import bitedu.bipa.simplesignbackend.validation.RestApiException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.websocket.Session;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -25,6 +28,9 @@ public class FormManageService {
 
     public List<FormAndCompDTO> selectFormAndCompList(FormAndCompDTO formAndCompDTO) {
         List<FormAndCompDTO> formAndCompList = formManageDAO.selectFormAndComp(formAndCompDTO);
+        if (!SessionUtils.hasIdAttribute("compId", Integer.parseInt(formAndCompDTO.getCompId()))) {
+            throw new RestApiException(CustomErrorCode.INACTIVE_USER);
+        }
         if(formAndCompList.size() < 1){
             throw new RestApiException(CommonErrorCode.RESOURCE_NOT_FOUND);
         }
@@ -35,6 +41,9 @@ public class FormManageService {
         FormDetailResDTO formDetail = new FormDetailResDTO();
         try{
             formDetail = formManageDAO.selectFormDetail(code);
+            if (!SessionUtils.hasIdAttribute("compId", formDetail.getCompId())) {
+                throw new RestApiException(CustomErrorCode.INACTIVE_USER);
+            }
             ArrayList<FormDetailScopeDTO> formDetailScopeList = formManageDAO.selectFormScope(code);
             formDetail.setScope(formDetailScopeList);
             if(formDetail.getCode() == 0){
@@ -71,7 +80,11 @@ public class FormManageService {
     @Transactional
     public Boolean formDetailRegist(FormDetailResDTO formDetail) {
         try {
+            if (!SessionUtils.hasIdAttribute("compId", formDetail.getCompId())) {
+                throw new RestApiException(CustomErrorCode.INACTIVE_USER);
+            }
             formManageDAO.insertFormDetail(formDetail);
+
             int formCode = commonMapper.getLastInsertId();
             if(formCode == 0){
                 throw new RestApiException(CommonErrorCode.UNEXPECTED_TYPE);
@@ -116,6 +129,9 @@ public class FormManageService {
     @Transactional
     public Boolean formDetailChange(FormDetailResDTO formDetail) {
         try {
+            if (!SessionUtils.hasIdAttribute("compId", formDetail.getCompId())) {
+                throw new RestApiException(CustomErrorCode.INACTIVE_USER);
+            }
             formManageDAO.updateFormDetail(formDetail);
             int formCode = formDetail.getCode();
             if(formCode == 0){
@@ -200,7 +216,12 @@ public class FormManageService {
     }
 
     public Boolean removeForm(int code) {
+        FormDetailResDTO formDetail = new FormDetailResDTO();
         try {
+            formDetail = formManageDAO.selectFormDetail(code);
+            if (!SessionUtils.hasIdAttribute("compId", formDetail.getCompId())) {
+                throw new RestApiException(CustomErrorCode.INACTIVE_USER);
+            }
             formManageDAO.deleteForm(code);
         }catch (Exception e){
             e.printStackTrace();
